Make the behavior of account ID based routing configurable in SDK (#4094)
## Motivation and Context
This is part 5, and the last, in the series of supporting account
ID-based routing (continuation of
https://github.com/smithy-lang/smithy-rs/pull/4047,
https://github.com/smithy-lang/smithy-rs/pull/4057,
https://github.com/smithy-lang/smithy-rs/pull/4087, and
https://github.com/smithy-lang/smithy-rs/pull/4091).

## Description
This PR makes the feature of account ID-based routing configurable
across the SDK. The `AccountIdEndpointMode` configuration takes one of
three values: `preferred`, `disabled`, or `required`, with `preferred`
being the default value (see [the reference
guide](https://docs.aws.amazon.com/sdkref/latest/guide/feature-account-endpoints.html)
for more information).

This PR makes the `AccountIdEndpointMode` configurable by:
- Exposing the setter(s) for `account_id_endpoint_mode` on both
[`ConfigLoader`](https://github.com/smithy-lang/smithy-rs/blob/a008dc282e6e9615064302616a81c36d43c6eb5b/aws/rust-runtime/aws-config/src/lib.rs#L631-L638)
and
[`SdkConfig`](https://github.com/smithy-lang/smithy-rs/blob/a008dc282e6e9615064302616a81c36d43c6eb5b/aws/rust-runtime/aws-types/src/sdk_config.rs#L181-L197).
- Passing the configured value through the existing configuration chain:
`ConfigLoader` → `SdkConfig` → service config's `Builder` →
`read_before_execution` method of the endpoint params' interceptors.

In addition, this PR includes the following supporting changes:
- Defines an enum `AccountIdEndpointMode` in the `aws-types` crate.
- Breaks APIs in the `dynamodb` to set
[`account_id`](https://docs.rs/aws-sdk-dynamodb/1.71.2/aws_sdk_dynamodb/config/struct.Builder.html#method.account_id)
and
[`account_id_endpoint_mode`](https://docs.rs/aws-sdk-dynamodb/1.71.2/aws_sdk_dynamodb/config/struct.Builder.html#method.account_id_endpoint_mode)
(introduced in smithy-rs#3792). This involves removing the setters for
`account_id` and updating the setters for `account_id_endpoint_mode` to
accept the `AccountIdEndpointMode` enum instead of a `String`.

## Testing
- Added tests for sourcing `AccountIdEndpointMode` from env config
- Passed `dynamodb`'s `endpoint_tests`, which cover the account ID mode
test cases in the design spec, with code changes in the PR
- Added codegen test to verify the account ID based routing with/without
auth

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._