Remove merge queue (#4124)
The merge queue was introduced when there were around 10 developers
working simultaneously in this repository. However, with the fewer
number of developers, the value of the queue has diminished. In some
cases, the wait time in the queue can delay the release of urgent fixes.

Note that we can re-enable the merge queue in the future if needed.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._