Fix bug with enum codegen
When the first enum generated has the @sensitive trait the opaque type
underlying the Unknown variant inherits that sensitivity. This means that
it does not derive Debug. Since the module is only generated once this
causes a problem for non-sensitive enums that rely on the type deriving
Debug so that they can also derive Debug.