fix h2 goaway error retry regression (#4145)
## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here -->
fixes https://github.com/awslabs/aws-sdk-rust/issues/1272

## Description
<!--- Describe your changes in detail -->

The hyper 1.x work introduced a regression to the hyper legacy 0.14
client's ability to retry h2::GoAway errors. This was due to the h2
version that each hyper version is pinned to as described in the
reproduction in the issue. This is similar to the issue we've seen in
the past with different major versions of a type being added to config
bag.

Also tested against hyper 1.x using the same mechanism supplied in the
[reproduction](https://github.com/awslabs/aws-sdk-rust/issues/1272#issuecomment-2887562827)
(see
https://github.com/aajtodd/hyper/commit/7fe4ee6b091dc37ed5b6cd17d1da067a42e3b6a4)
to verify it also works as intended.

## Testing
Tested repro locally with and without the fix. Without the fix I can see
no retries attempted and the [warning
message](https://github.com/smithy-lang/smithy-rs/blob/release-2025-05-19/rust-runtime/aws-smithy-http-client/src/hyper_legacy.rs#L393).
With the fix retries are attempted and no warning message.

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [x] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.
----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._