Revert version 2.1 of the IMDS credentials provider (#4187)
## Motivation and Context
We released the feature [in this
release](https://github.com/awslabs/aws-sdk-rust/releases/tag/release-2025-06-12),
but till the infrastructure is fully ready, we'll temporarily revert the
functionality.

## Testing
- CI

## Checklist
- [x] For changes to the AWS SDK, generated SDK code, or SDK runtime
crates, I have created a changelog entry Markdown file in the
`.changelog` directory, specifying "aws-sdk-rust" in the `applies_to`
key.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._