Http-1x update phase 1 (#4181)
## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here -->
The first steps to defaulting to `http-1x` dependencies instead of
`http-02x`

## Description
<!--- Describe your changes in detail -->
For this PR I updated the dependencies on SDK crates to enable their
`http-1x` features and worked backwards from there fixing the things
that broke (this would be easier to see if the diff was working but it
seems to be broken for feature branches after the new ECR update). This
involved:
* Updating the `http_request_checksum` inlineable to use `http-1x` types
(and removing the `http-1x` feature since it is no longer needed and
wasn't a "real" feature to begin with
* Creating a new `pub mod content_encoding_http_1x` module in the
`aws-runtime` crate. Most logic here is dedicated to implementing the
`http_body_1x::Body` trait for `AwsChunkedBody`. I initially considered
keeping this in the existing `content_encoding` module, but with the
current feature flags and all of the changes I had to make it felt
cleaner to break it out.
* Update body polling logic for `ChecksumBody` calculation and
validation for `http-1x` types


## Testing
<!--- Please describe in detail how you tested your changes -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
Updated existing tests and added new ones for `http-1x` types

## TODOs
Future PRs will
* Update code generation logic in codegen-core to use `http-1x`
* Remove pre-1.0 `http` dependencies (anywhere they aren't re-exported)
* Normalize dependency names to `http-1x`, `http-02x`, etc

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._