Make auth scheme and auth sheme option resolver configurable (#4182)
## Motivation and Context
The third PR in the auth alignment series (follow-up on
https://github.com/smithy-lang/smithy-rs/pull/4161 and
https://github.com/smithy-lang/smithy-rs/pull/4175)

## Description
This PR allows users to set custom auth scheme and auth scheme option
resolver on service config builder. Their setters are on the service
config builder rather than `ConfigLoader` or `SdkConfig`, because some
of them take as input service-specific auth scheme `Params` (so makes
sense for the others to be on the service config builder for
consistency).

Regarding removing `optionalAuth` transformer from S3 model
([discussion](https://github.com/smithy-lang/smithy-rs/pull/4175#discussion_r2145177739)),
~this PR backs up the implementation of
[no_credentials](https://docs.rs/aws-config/latest/aws_config/struct.ConfigLoader.html#method.no_credentials)
to compensate for the removal. However, whether we ultimately remove
that S3 transformer or not will be assessed and finalized in a
subsequent PR.~ this PR has restored it with a tweak that a special
synthesized trait is added to the S3's service shape. This is
in-contrast to applying `@optionalAuth` to every operation as previously
done, which no longer works with the logic of `AuthIndex` to determine a
list of operation overrides for auth scheme option.

Also removed `UnsignedPayloadTrait` in determining operation overrides
for the default auth scheme resolver
([discussion](https://github.com/smithy-lang/smithy-rs/pull/4161#discussion_r2136339891))

## Testing
- CI
- Enabled all tests disabled in the [previous
PR](https://github.com/smithy-lang/smithy-rs/pull/4175)
- Added codegen tests to override auth scheme and auth scheme option
resolver both at service config level and at operation level

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._