Update SdkBody to support http-1x bodies (#4199)
## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here -->
Continuing the work to default to `http-1x` types under the hood. 

## Description
<!--- Describe your changes in detail -->
This is unfortunately an enormous PR (at least in terms of files
touched, not total lines). The vast majority of changes are just
replacing references to `http_02x` with `http_1x` or `http_body_04x`
with `http_body_1x` and updating code or feature flags to match.

The primary functional changes are in the vicintiy of
`rust-runtime/aws-smithy-types/src/body.rs`, updating methods for
`SdkBody` to understand `http_1x` types. This required several changes:
* Add a new `HttpBody1` variant to the `enum BoxBody`
* Update `SdkBody::from_body_1_x` to construct a `HttpBody1` variant
instead of the `HttpBody04` variant it was previously
* Add a new `trailers: Option<VecDeque<http_1x::HeaderMap>>` field to
`SdkBody` to cache polled trailers (`http_body_1x` doesn't allow polling
trailers separately, but `SdkBody` does, so we have to cache at least
the first trailer frame if it is encountered while polling the body)
* Update `SdkBody::{poll_next, poll_next_trailers}` to work with the new
`HttpBody1x` variant.

## Testing
<!--- Please describe in detail how you tested your changes -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
Updated most existing tests to use `http-1x` types (Although I have
almost certainly missed some)

Note that some tests are failing here and that is expected:
* Semver tests: we have changed/deleted some `http_02x` trait impls in
pre-1.0 crates, so we expect these to fail
* Server tests: I haven't touched these and will work with Fahad to get
them updated

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._