Adjust `no_auth` test expectation to address semver hazards failure (#4211)
## Motivation and Context
This PR is a prerequisite for #4203.

# Details
PR4203 has updated some of the `AuthSchemeId` string values, which
causes the `no_auth` test to fail in the semver hazards check. This
happens because the inner string value for `NO_AUTH_SCHEME_ID` has been
modified from `no_auth` to `noAuth`.

To resolve this, this PR updates the test expectation so that it no
longer depends on raw string values. Once this PR is merged and released
from `aws-sdk-rust`, PR4203 should pass the semver hazards check
successfully.

Also the test has been moved to `no_auth.rs` since it is not specific to
S3 Express; it was originally placed in `express.rs` because the issue
was first discovered while investigating S3 Express behavior.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._