Fix broken workflows since the introduction of private ecr repository (#4220)
## Motivation and Context
Minor cleanup following #4153

## Description
The following workflows have been broken due to minor issues—such as
missing secrets, not using the latest upload scripts, and so on:
- `pull-request-updating-lockfiles.yml`
- `update-lockfiles.yml` and `manual-update-lockfiles.yml` are callers
of the given workflow
- `ci-main.yml`
- `dry-run-release-scheduled.yml`

## Testing
The changes to `ci-main.yml` and `dry-run-release-scheduled.yml` haven't
been verified, as these workflows only run in live environments (if they
are still broken after this PR, will look into them further).

For the changes to `pull-request-updating-lockfiles.yml`, manually
triggered `manual-update-lockfiles.yml` and confirmed that #4219 has
been created.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._