Initial setup for credential resolution feature tracking (#4224)
## Motivation and Context

Initial setup for tracking credential/identity resolution based
features.

## Description

The biggest change here is the addition of a new `aws-features` crate
containing the `sdk_features` module that used to live in `aws-runtime`.
This was done so that the `AwsSdkFeature` type could be referenced in
the `aws-credential-types` crate without causing a circular dependency.

Initially I was just going to pass along the types without downcasting,
but they would have had to be downcast eventually, since in the
`Credentials` typemap they are `Vec<AwsSdkFeature>` but to add them to a
`Layer` we need to extract the individual `AwsSdkFeature`s from the
`Vec`. This could either happen in `aws-credential-types` where
`From<Credentials> for Identity` is implemented or in
`aws-smithy-runtime` in the orchestrator where the `Identity` is
resolved. Since these credentials types are AWS specific it seemed to
make more sense to keep it in `aws/rust-runtime`.

Other changes:
* Add a typemap to `Credentials` (and manually implement the previously
derived traits for it) to carry the feature information
* Update the `From<Credentials> for Identity` implementation to extract
the `AwsSdkFeatures` and pass them in a `Layer` to the `Identity`'s
typemap.
* Update `resolve_identity` in the orchestrator to extract the
`FrozenLayer` from `Identity` and insert it in the `ConfigBag` so that
the `UserAgentInterceptor` can extract it later.

## Testing
<!--- Please describe in detail how you tested your changes -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
Added new tests around `Credentials` equality and the updated
`From<Credentials> for Identity` implementation.

**Note on failing semver test:** Failing because the `UnwindSafe` traits
are no longer auto implemented for `Credentials`

```
--- failure auto_trait_impl_removed: auto trait no longer implemented ---

Description:
A public type has stopped implementing one or more auto traits. This can break downstream code that depends on the traits being implemented.
        ref: https://doc.rust-lang.org/reference/special-types-and-traits.html#auto-traits
       impl: https://github.com/obi1kenobi/cargo-semver-checks/tree/v0.41.0/src/lints/auto_trait_impl_removed.ron

Failed in:
  type Credentials is no longer UnwindSafe, in /home/build/workspace/smithy-rs/aws-sdk/sdk/aws-credential-types/src/credentials_impl.rs:29
  type Credentials is no longer RefUnwindSafe, in /home/build/workspace/smithy-rs/aws-sdk/sdk/aws-credential-types/src/credentials_impl.rs:29Cloning base
```

This is due to us adding a `HashMap<TypeId, TypeErasedBox>` to
`Credentials`. We could likely wrap this in an `Arc<Mutex<>>` to keep it
mutable and get back the unwind safety, but that doesn't feel like it
justifies the added complexity. Open to debate on this one though.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._