Consolidating some layers in Dockerfile (#4251)
## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here -->
We recently started seeing errors with the Acquire Base Image actions
([ex](https://github.com/smithy-lang/smithy-rs/actions/runs/16784733761/job/47532356543)).
These were all caused by:
```
------STDERR---------
failed to register layer: max depth exceeded
-------------------
```

We are not quite sure why these started showing up in the past two days.
The last change to our Dockerfile was about two weeks ago in
https://github.com/smithy-lang/smithy-rs/pull/4217. The issue might be
related to the clone of smithy-rs introduced in that change, but that
didn't cause any issues at the time.

To get around this and unblock CI until we root cause it this PR reduces
the number of layers in the image by grouping the ARGs and the RUNs for
the local tool installation into a single statement each.


## Description
<!--- Describe your changes in detail -->

## Testing
<!--- Please describe in detail how you tested your changes -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._