pin crc-fast to <1.4 to workaround SIGILL (#4264)
## Description

[Canary](https://github.com/awslabs/aws-sdk-rust/actions/runs/16923899511/job/47955897284)
is currently failing and we have some customers reporting on this
[issue](https://github.com/awesomized/crc-fast-rust/issues/14) that
1.4.0 is causing SIGILL for them. Pinning crc-fast to 1.3.0 temporarily
while this gets sorted.



## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [x] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.
- [x] For changes to the AWS SDK, generated SDK code, or SDK runtime
crates, I have created a changelog entry Markdown file in the
`.changelog` directory, specifying "aws-sdk-rust" in the `applies_to`
key.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._