Add `Clone` to `HickoryDnsResolver` and `TokioDnsResolver` (#4289)
## Description
<!--- Describe your changes in detail -->
The `HickoryDnsResolver` and `TokioDnsResolver` were not `Clone` making
it impossible to use them in the http_client builder's
`build_with_resolver` method.

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [x] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.


----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._